Refactor: move most of sock into modules
