+++ server/src/lib.rs
@@ -1,0 +1,3 @@
+pub trait ParseFrom<T, Out = Self> {
+    fn parse (from: T) -> Result<Out, String>;
+}
